lukas_mandik_photos/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.54.0
django-dump-load-utf8==0.0.4
django-extensions==3.2.1
django-livereload-server==0.4
djangorestframework==3.14.0
exif==1.6.0
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.4.0
plum-py==0.8.5
pycparser==2.21
pyexifinfo==0.4.0
PyExifTool==0.5.5
PyJWT==2.6.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
watchdog==3.0.0
wsproto==1.2.0
